White label about modal. Move strings to config.
