+++ assets/base/config.json
@@ -1,9 +1,14 @@
 {
+    "AppName": "Mattermost",
     "DefaultServerUrl": "",
     "TestServerUrl": "http://localhost:8065",
     "DefaultTheme": "default",
     "ShowErrorsList": false,
     "MinServerVersion": "4.0.0",
+    "EELearnURL": "about.mattermost.com",
+    "TeamEditionLearnURL": "mattermost.org",
+    "AboutTeamURL": "http://www.mattermost.org/",
+    "AboutEnterpriseURL": "http://about.mattermost.com/",
     "PlatformNoticeURL": "https://about.mattermost.com/platform-notice-txt/",
     "MobileNoticeURL": "https://about.mattermost.com/mobile-notice-txt/",
     "SegmentApiKey": "3MT7rAoC0OP7yy3ThzqFSAtKzmzqtUPX",
